+++ Core/DomainModel/PurchaseOperation/PurchaseReceivalDetail.cs
@@ -11,9 +11,9 @@ namespace Core.DomainModel
         public string Code { get; set; }
         public int PurchaseReceivalId { get; set; }
         public int ItemId { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public bool IsAllInvoiced { get; set; }
-        public int PendingInvoicedQuantity { get; set; }
+        public decimal PendingInvoicedQuantity { get; set; }
         public int PurchaseOrderDetailId { get; set; }
 
         public decimal COGS { get; set; }
