+++ Core/Interface/Service/PurchaseOperation/IPurchaseOrderDetailService.cs
@@ -25,7 +25,7 @@ namespace Core.Interface.Service
         PurchaseOrderDetail UnconfirmObject(PurchaseOrderDetail purchaseOrderDetail, IPurchaseReceivalDetailService _purchaseReceivalDetailService,
                                            IStockMutationService _stockMutationService, IItemService _itemService, IBlanketService _blanketService,
                                            IWarehouseItemService _warehouseItemService);
-        PurchaseOrderDetail SetReceivalComplete(PurchaseOrderDetail purchaseOrderDetail, int Quantity);
-        PurchaseOrderDetail UnsetReceivalComplete(PurchaseOrderDetail purchaseOrderDetail, int Quantity, IPurchaseOrderService _purchaseOrderService);
+        PurchaseOrderDetail SetReceivalComplete(PurchaseOrderDetail purchaseOrderDetail, decimal Quantity);
+        PurchaseOrderDetail UnsetReceivalComplete(PurchaseOrderDetail purchaseOrderDetail, decimal Quantity, IPurchaseOrderService _purchaseOrderService);
    }
 }
