+++ Service/Service/SalesOperation/SalesOrderDetailService.cs
@@ -109,7 +109,7 @@ namespace Service.Service
             return salesOrderDetail;
         }
 
-        public SalesOrderDetail SetDeliveryComplete(SalesOrderDetail salesOrderDetail, int Quantity)
+        public SalesOrderDetail SetDeliveryComplete(SalesOrderDetail salesOrderDetail, decimal Quantity)
         {
             salesOrderDetail.PendingDeliveryQuantity -= Quantity;
             if (salesOrderDetail.PendingDeliveryQuantity == 0) { salesOrderDetail.IsAllDelivered = true; }
@@ -117,7 +117,7 @@ namespace Service.Service
             return salesOrderDetail;
         }
 
-        public SalesOrderDetail UnsetDeliveryComplete(SalesOrderDetail salesOrderDetail, int Quantity, ISalesOrderService _salesOrderService)
+        public SalesOrderDetail UnsetDeliveryComplete(SalesOrderDetail salesOrderDetail, decimal Quantity, ISalesOrderService _salesOrderService)
         {
             SalesOrder salesOrder = _salesOrderService.GetObjectById(salesOrderDetail.SalesOrderId);
             _salesOrderService.UnsetDeliveryComplete(salesOrder);
